Introduce small string optimization for big number
Since the default fibdrv cannot support the value after the 92th
term of the fibonacci in the type long long. As result, we use string
to store the values and use the string optimization for the reservation
of heap furthermore. With values stored in the format of strings, we can
use system call "read" to write back the values of fibonacci sequence and
use the value returned in 'int' to represent the digits of that number.