Introduce the fast doubling method into Fibonacci
The fast doubling method is added in fibdrv.c to make different
ways to evaluate the fibonacci sequence. The difference between fast
doubling method and dynamic programming method is that the time
complexity of dynamic programming one is O(n) while that of the fast
doubling one is O(log n). However, dynamic programming method only
needs addition in computation while fast doubling method needs not only
addition but also substraction and even muliplication. So, the
computation should be implemented because we use the special data
structure "xs" to store the huge numbers. And it is worth doing an
experiment to discuss the performance of both methods.