Fix some problem about memory leak
There are some operations in the previous patch that make a non-null
pointer point to another space, which may cause the memory leak. To
fix the problem, we free the previous space before a new assignment
to a pointer and free the space if it has not been used anymore. Also,
we delete the printks in the code that were used to debug to make the
performance better and decrease those unnecessary messages.